Bug fix no delete das viagens

+++ server/services/viagem/delete.js
@@ -2,8 +2,23 @@ import { _req, _db, _val, _out } from "@netuno/server-types";
 
 const uid = _req.getString("uid");
 
-if (_db.delete("viagem", uid)) {
-    _out.json(_val.map().set("sucesso", true));
+const dbViagem = _db.query("SELECT id FROM viagem WHERE uid = ?", uid);
+
+if (dbViagem.length > 0) {
+    const viagemId = dbViagem[0].getInt("id");
+
+    try {
+        _db.execute("DELETE FROM despesa WHERE viagem_id = ?", viagemId);
+        _db.execute("DELETE FROM membro WHERE viagem_id = ?", viagemId);
+
+        if (_db.delete("viagem", uid)) {
+            _out.json(_val.map().set("sucesso", true));
+        } else {
+            _out.json(_val.map().set("sucesso", false).set("erro", "Falha ao apagar a viagem."));
+        }
+    } catch (e) {
+        _out.json(_val.map().set("sucesso", false).set("erro", "Erro interno ao limpar os dados da viagem."));
+    }
 } else {
-    _out.json(_val.map().set("sucesso", false).set("erro", "Apaga primeiro os gastos associados a esta viagem."));
+    _out.json(_val.map().set("sucesso", false).set("erro", "Viagem não encontrada."));
 }
